Update notifications fetch services to use the session
https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have multiple journeys for notifications.

This change updates the notification setup fetch services to use the session.

This will enable us to use the journey context within the query to handle ad-hoc returns and other notification types.

This change for ad-hoc will come at in a later change.
