+++ app/services/notifications/setup/download-recipients.service.js
@@ -5,11 +5,9 @@
  * @module DownloadRecipientsService
  */
 
-const DetermineReturnsPeriodService = require('./determine-returns-period.service.js')
 const DownloadRecipientsPresenter = require('../../../presenters/notifications/setup/download-recipients.presenter.js')
 const FetchDownloadRecipientsService = require('./fetch-download-recipients.service.js')
 const SessionModel = require('../../../models/session.model.js')
-const { transformStringOfLicencesToArray } = require('../../../lib/general.lib.js')
 
 /**
  * Orchestrates fetching and formatting the data needed for the notifications setup download link
@@ -23,17 +21,10 @@ const { transformStringOfLicencesToArray } = require('../../../lib/general.lib.j
  */
 async function go(sessionId) {
   const session = await SessionModel.query().findById(sessionId)
-  const { notificationType, referenceCode, returnsPeriod } = session
 
-  const determinedReturnsPeriod = DetermineReturnsPeriodService.go(returnsPeriod)
+  const { notificationType, referenceCode } = session
 
-  const removeLicences = transformStringOfLicencesToArray(session.removeLicences)
-
-  const recipients = await FetchDownloadRecipientsService.go(
-    determinedReturnsPeriod.returnsPeriod.dueDate,
-    determinedReturnsPeriod.summer,
-    removeLicences
-  )
+  const recipients = await FetchDownloadRecipientsService.go(session)
 
   const formattedData = DownloadRecipientsPresenter.go(recipients)
