+++ app/services/notifications/setup/fetch-download-recipients.service.js
@@ -6,6 +6,8 @@
  */
 
 const { db } = require('../../../../db/db.js')
+const DetermineReturnsPeriodService = require('./determine-returns-period.service.js')
+const { transformStringOfLicencesToArray } = require('../../../lib/general.lib.js')
 
 /**
  * Formats the contact data from which recipients will be determined for the `/notifications/setup/download` link
@@ -43,14 +45,16 @@ const { db } = require('../../../../db/db.js')
  * received from this query (For either registered to unregistered licence). We expect to see duplicate licences with
  * different contacts types (but still preferring the registered over unregistered licence).
  *
- * @param {string} dueDate - The 'due' date for outstanding return logs to fetch contacts for
- * @param {string} summer - Whether we are looking for outstanding summer or all year return logs
- * @param {string[]} removeLicences - The licences to exclude from the results
+ * @param {module:SessionModel} session - The session instance to format
  *
  * @returns {Promise<object[]>} - matching recipients
  */
-async function go(dueDate, summer, removeLicences) {
-  const { rows } = await _fetch(dueDate, summer, removeLicences)
+async function go(session) {
+  const { returnsPeriod, summer } = DetermineReturnsPeriodService.go(session.returnsPeriod)
+
+  const removeLicences = transformStringOfLicencesToArray(session.removeLicences)
+
+  const { rows } = await _fetch(returnsPeriod.dueDate, summer, removeLicences)
 
   return rows
 }
