+++ test/services/notifications/setup/download-recipients.service.test.js
@@ -17,15 +17,12 @@ const DownloadRecipientsService = require('../../../../app/services/notification
 
 describe('Notifications Setup - Download recipients service', () => {
   const referenceCode = 'RINV-00R1MQ'
-  const year = 2025
 
-  let clock
   let removeLicences
   let session
   let testRecipients
 
   before(async () => {
-    clock = Sinon.useFakeTimers(new Date(`${year}-01-01`))
     removeLicences = ''
 
     session = await SessionHelper.add({
@@ -34,10 +31,6 @@ describe('Notifications Setup - Download recipients service', () => {
 
     testRecipients = _recipients()
     Sinon.stub(FetchDownloadRecipientsService, 'go').resolves(testRecipients)
-  })
-
-  after(() => {
-    clock.restore()
   })
 
   it('correctly returns the csv string, filename and type', async () => {
